+++ aiobungie/interfaces/factory.py
@@ -39,6 +39,7 @@ if typing.TYPE_CHECKING:
     from aiobungie.crate import friends
     from aiobungie.crate import milestones
     from aiobungie.crate import profile
+    from aiobungie.crate import season
     from aiobungie.crate import user
     from aiobungie.internal import helpers
     from aiobungie.internal import traits
@@ -54,6 +55,7 @@ class FactoryInterface(abc.ABC):
 
     # Users, Memberships.
 
+    @abc.abstractmethod
     def deserialize_user(self, data: helpers.JsonObject) -> user.User:
         """Deserialize a raw JSON hard linked Bungie.net user payload into a user object.
 
@@ -70,6 +72,7 @@ class FactoryInterface(abc.ABC):
             A user object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_bungie_user(self, data: helpers.JsonObject) -> user.BungieUser:
         """Deserialize a raw JSON Bungie.net user only payload into a user object.
 
@@ -87,6 +90,7 @@ class FactoryInterface(abc.ABC):
             A Bungie user object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deseialize_found_users(
         self, payload: helpers.JsonObject
     ) -> typing.Sequence[user.DestinyUser]:
@@ -103,6 +107,7 @@ class FactoryInterface(abc.ABC):
             A Destiny user object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_partial_bungie_user(
         self, payload: helpers.JsonObject, *, noeq: bool = False
     ) -> user.PartialBungieUser:
@@ -128,6 +133,7 @@ class FactoryInterface(abc.ABC):
             A partial bungie user object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_destiny_user(
         self, payload: helpers.JsonObject, *, noeq: bool = False
     ) -> user.DestinyUser:
@@ -156,6 +162,7 @@ class FactoryInterface(abc.ABC):
         """
 
     # Deserialize a list of `destinyUserInfo`
+    @abc.abstractmethod
     def deserialize_destiny_members(
         self,
         data: typing.Union[helpers.JsonObject, helpers.JsonArray],
@@ -184,6 +191,7 @@ class FactoryInterface(abc.ABC):
             A sequence of destiny membership/user object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_user_themes(
         self, payload: helpers.JsonArray
     ) -> typing.Sequence[user.UserThemes]:
@@ -200,6 +208,7 @@ class FactoryInterface(abc.ABC):
             A sequence of bungie user themes object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_player(
         self, payload: helpers.JsonArray, /
     ) -> typing.Sequence[user.DestinyUser]:
@@ -222,6 +231,7 @@ class FactoryInterface(abc.ABC):
 
     # Clans, Groups.
 
+    @abc.abstractmethod
     def deseialize_clan_owner(self, data: helpers.JsonObject) -> clans.ClanMember:
         """Deserialize a raw JSON payload of clan founder information.
 
@@ -236,6 +246,7 @@ class FactoryInterface(abc.ABC):
             A clan owner object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_clan(
         self, payload: helpers.JsonObject, *, bound: bool = False
     ) -> clans.Clan:
@@ -263,6 +274,7 @@ class FactoryInterface(abc.ABC):
             A clan owner object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_group_member(
         self, payload: helpers.JsonObject
     ) -> helpers.NoneOr[clans.GroupMember]:
@@ -279,6 +291,7 @@ class FactoryInterface(abc.ABC):
             A group member object of the deserialized payload. This can return `None` if nothing was found.
         """
 
+    @abc.abstractmethod
     def deserialize_clan_admins(
         self, payload: helpers.JsonObject
     ) -> typing.Sequence[clans.ClanAdmin]:
@@ -295,6 +308,7 @@ class FactoryInterface(abc.ABC):
             A sequence of clan admins object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_clan_member(self, data: helpers.JsonObject, /) -> clans.ClanMember:
         """Deserialize a JSON payload of a clan member information.
 
@@ -309,6 +323,7 @@ class FactoryInterface(abc.ABC):
             A clan member object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_clan_members(
         self, data: helpers.JsonObject, /
     ) -> typing.Sequence[clans.ClanMember]:
@@ -325,6 +340,7 @@ class FactoryInterface(abc.ABC):
             A sequence of clan members of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_clan_convos(
         self, payload: helpers.JsonArray
     ) -> typing.Sequence[clans.ClanConversation]:
@@ -341,6 +357,7 @@ class FactoryInterface(abc.ABC):
             A sequence of clan conversations of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_clan_banners(
         self, payload: helpers.JsonObject
     ) -> typing.Sequence[clans.ClanBanner]:
@@ -359,6 +376,7 @@ class FactoryInterface(abc.ABC):
 
     # Application
 
+    @abc.abstractmethod
     def deserialize_app_owner(
         self, payload: helpers.JsonObject
     ) -> application.ApplicationOwner:
@@ -375,6 +393,7 @@ class FactoryInterface(abc.ABC):
             An application owner object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_app(self, payload: helpers.JsonObject) -> application.Application:
         """Deserialize a JSON payload of Bungie Developer portal application information.
 
@@ -391,6 +410,7 @@ class FactoryInterface(abc.ABC):
 
     # Characters.
 
+    @abc.abstractmethod
     def deserialize_character(
         self, payload: helpers.JsonObject
     ) -> typing.Optional[character.Character]:
@@ -409,6 +429,7 @@ class FactoryInterface(abc.ABC):
 
     # Profiles.
 
+    @abc.abstractmethod
     def deserialize_profile(
         self, payload: helpers.JsonObject, /
     ) -> typing.Optional[profile.Profile]:
@@ -421,10 +442,11 @@ class FactoryInterface(abc.ABC):
 
         Returns
         -------
-        `aiobungie.crate.Profile`
+        `typing.Optional[aiobungie.crate.Profile]`
             A profile object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_linked_profiles(
         self, payload: helpers.JsonObject
     ) -> profile.LinkedProfile:
@@ -462,6 +484,7 @@ class FactoryInterface(abc.ABC):
 
     # Inventory entities and Definitions.
 
+    @abc.abstractmethod
     def deserialize_inventory_entity(
         self, payload: helpers.JsonObject, /
     ) -> entity.InventoryEntity:
@@ -482,6 +505,7 @@ class FactoryInterface(abc.ABC):
 
     # Activities.
 
+    @abc.abstractmethod
     def deserialize_activity(self, payload: helpers.JsonObject, /) -> activity.Activity:
         """Deserialize a JSON payload of an occurred activity information.
 
@@ -498,6 +522,7 @@ class FactoryInterface(abc.ABC):
 
     # Milestones.
 
+    @abc.abstractmethod
     def deserialize_public_milestone_content(
         self, payload: helpers.JsonObject
     ) -> milestones.Milestone:
@@ -516,6 +541,7 @@ class FactoryInterface(abc.ABC):
 
     # Social and friends.
 
+    @abc.abstractmethod
     def deserialize_friend(self, payload: helpers.JsonObject, /) -> friends.Friend:
         """Deserialize a JSON payload of a Bungie friend information.
 
@@ -530,6 +556,7 @@ class FactoryInterface(abc.ABC):
             A friend object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_friends(
         self, payload: helpers.JsonObject
     ) -> typing.Sequence[friends.Friend]:
@@ -548,6 +575,7 @@ class FactoryInterface(abc.ABC):
             A sequence of friends object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_friend_requests(
         self, payload: helpers.JsonObject
     ) -> friends.FriendRequestView:
@@ -568,6 +596,7 @@ class FactoryInterface(abc.ABC):
 
     # Fireteams.
 
+    @abc.abstractmethod
     def deserialize_fireteams(
         self, payload: helpers.JsonObject
     ) -> helpers.NoneOr[typing.Sequence[fireteams.Fireteam]]:
@@ -584,6 +613,7 @@ class FactoryInterface(abc.ABC):
             A sequence of fireteam object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_fireteam_destiny_users(
         self, payload: helpers.JsonObject
     ) -> fireteams.FireteamUser:
@@ -600,6 +630,7 @@ class FactoryInterface(abc.ABC):
             A fireteam user object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_fireteam_members(
         self, payload: helpers.JsonObject, *, alternatives: bool = False
     ) -> typing.Optional[typing.Sequence[fireteams.FireteamMember]]:
@@ -619,6 +650,7 @@ class FactoryInterface(abc.ABC):
             An optional sequence of the fireteam members object of the deserialized payload.
         """
 
+    @abc.abstractmethod
     def deserialize_available_fireteams(
         self, data: helpers.JsonObject, *, no_results: bool = False
     ) -> typing.Union[
@@ -637,4 +669,38 @@ class FactoryInterface(abc.ABC):
         -------
         `typing.Union[aiobungie.crate.fireteams.AvalaibleFireteam, typing.Sequence[aiobungie.crate.fireteams.AvalaibleFireteam]]` # noqa: E501
             An available fireteam or a sequence of available fireteam object of the deserialized payload.
+        """
+
+    @abc.abstractmethod
+    def deserialize_seasonal_artifact(
+        self, payload: helpers.JsonObject
+    ) -> season.Artifact:
+        """Deserialize a JSON payload of a Destiny 2 seasonal artifact information.
+
+        Parameters
+        ----------
+        payload : `aiobungie.internal.helpers.JsonObject`
+            The JSON payload.
+
+        Returns
+        -------
+        `aiobungie.crate.Artifact`
+            A seasonal artifact object of the deserialized payload.
+        """
+
+    @abc.abstractmethod
+    def deserialize_profile_progression(
+        self, payload: helpers.JsonObject
+    ) -> profile.ProfileProgression:
+        """Deserialize a JSON payload of a profile progression component.
+
+        Parameters
+        ----------
+        payload : `aiobungie.internal.helpers.JsonObject`
+            The JSON payload.
+
+        Returns
+        -------
+        `aiobungie.crate.ProfileProgression`
+            A profile progression component object of the deserialized payload.
         """